refactor(ai): 统一创建接口返回受影响行数

+++ lucky-modules/lucky-ai/src/main/java/com/lucky/ai/service/IAiApiKeyService.java
@@ -22,7 +22,7 @@ public interface IAiApiKeyService {
      * @param query 创建请求vo
      * @return 结果
      */
-    Long insertApiKey(AiApiKeySaveQuery query);
+    int insertApiKey(AiApiKeySaveQuery query);
 
     /**
      * 更新 API 密钥
